Rename: session.removeCookie() -> session.destroyCookie()
