+++ server/api/session/destroyCookie/controller.js
@@ -2,14 +2,14 @@ import compose from 'koa-compose';
 import Boom from 'boom';
 import { decorateSession, isUserAuthenticated } from '../../../middleware/auth';
 import packJSONRPC from '../../../middleware/packJSONRPC';
-import { removeSessionCookie } from './service';
+import { destroySessionCookie } from './service';
 
 async function handler(ctx) {
   const { response, cookies } = ctx;
 
-  const isSessionCookieRemoved = await removeSessionCookie(ctx, cookies.get('session'));
+  const isSessionCookieDestroyed = await destroySessionCookie(ctx, cookies.get('session'));
 
-  if (!isSessionCookieRemoved) {
+  if (!isSessionCookieDestroyed) {
     throw Boom.internal();
   }
