+++ server/api/session/destroyCookie/index.js
