+++ server/api/session/index.js
@@ -3,7 +3,7 @@ import createSession from './create';
 import destroySession from './destroy';
 import refreshSession from './refresh';
 import setSessionCookie from './setCookie';
-import removeSessionCookie from './removeCookie';
+import destroySessionCookie from './destroyCookie';
 
 const router = Router();
 
@@ -13,6 +13,6 @@ router.post('session.create', '/session.create', createSession);
 router.post('session.destroy', '/session.destroy', destroySession);
 router.post('session.refresh', '/session.refresh', refreshSession);
 router.post('session.setCookie', '/session.setCookie', setSessionCookie);
-router.post('session.removeCookie', '/session.removeCookie', removeSessionCookie);
+router.post('session.destroyCookie', '/session.destroyCookie', destroySessionCookie);
 
 export default router;
